project completed
This Graph Coloring Project contains an algorithm to color a given graph with the least number of colors. The graph coloring problem, a widely used problem in graph theory, is about assigning different colors to the nodes of a graph to ensure that neighboring nodes do not have the same color. This project presents a greedy algorithm to solve the graph coloring problem.